refactor(record) : 데이터 타입 수정

+++ src/main/java/site/coach_coach/coach_coach_server/recorddata/dto/RecordDataRequest.java
@@ -4,9 +4,9 @@ import lombok.Builder;
 
 @Builder
 public record RecordDataRequest(
-	Integer weight,
-	Integer skeletalMuscle,
-	Integer fatPercentage,
+	Double weight,
+	Double skeletalMuscle,
+	Double fatPercentage,
 	Double bmi
 ) {
 }
